requirements.txt
jinja2>=3.1.0
click>=8.0.0
pyyaml>=6.0
pandas>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=23.0.0
rich>=13.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
